pyproject.toml
[project]
name = "lpitPublisher"
version = "0.0.1"
description = "A tool to publish LPiT documents."
authors = [
    {name = "Stephen Gaito", email = "stephen@perceptisys.co.uk"},
]
dependencies = [
    "pyyaml>=6.0",
    "types-PyYAML>=6.0.12.20250915",
    "jinja2>=3.1.6",
]
requires-python = ">=3.10"
license = {text = "Apache2.0"}

[project.scripts]
lpitCollector     = "lpitPublisher.collector:cli"
lpitGenerator     = "lpitPublisher.generator:cli"
lpitDocumentInit  = "lpitPublisher.documentInit:cli"
lpitStylesUpdater = "lpitPublisher.stylesUpdater:cli"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.mypy]
python_executable = "./.venv/bin/python"

